Add Azure Functions project. Compiles, but doesn't work because cannot find CodeAnalysis asm at runtime (though basically identical to Web).
